Create wrapper for AudioFocusRequest
