+++ audioswitch/src/main/java/com/twilio/audioswitch/AudioFocusRequestWrapper.kt
@@ -1,0 +1,22 @@
+package com.twilio.audioswitch
+
+import android.annotation.SuppressLint
+import android.media.AudioAttributes
+import android.media.AudioFocusRequest
+import android.media.AudioManager
+
+class AudioFocusRequestWrapper {
+
+    @SuppressLint("NewApi")
+    fun buildRequest(): AudioFocusRequest {
+        val playbackAttributes = AudioAttributes.Builder()
+                .setUsage(AudioAttributes.USAGE_VOICE_COMMUNICATION)
+                .setContentType(AudioAttributes.CONTENT_TYPE_SPEECH)
+                .build()
+        return AudioFocusRequest.Builder(AudioManager.AUDIOFOCUS_GAIN_TRANSIENT)
+                .setAudioAttributes(playbackAttributes)
+                .setAcceptsDelayedFocusGain(true)
+                .setOnAudioFocusChangeListener { i: Int -> }
+                .build()
+    }
+}
